campaigns/shai-hulud-2025/requirements.txt
# Cyber Threat Video Pipeline Dependencies
# Install: pip install -r requirements.txt

# Google Gemini
google-generativeai>=0.8.0

# OpenAI (Sora 2)
openai>=1.50.0

# ElevenLabs TTS
elevenlabs>=1.0.0

# HTTP & async
httpx>=0.27.0
aiohttp>=3.9.0

# CLI & utilities
click>=8.1.0
rich>=13.0.0
python-dotenv>=1.0.0

# Data processing
pydantic>=2.0.0pytest
